fix: create the health checker in `access` phase
Allow to call yield function in the healthcheck.


Fix #1851
Fix #2842
